WorkflowDescriptorSpec
