+++ engine/src/test/scala/cromwell/engine/workflow/lifecycle/WorkflowExecutionActorSpec.scala
@@ -5,7 +5,8 @@ import akka.testkit.{TestActorRef, EventFilter}
 import com.typesafe.config.ConfigFactory
 import cromwell.CromwellTestkitSpec
 import cromwell.core.WorkflowId
-import cromwell.engine.backend.{BackendConfigurationEntry, CromwellBackends, WorkflowDescriptorBuilder}
+import cromwell.engine.backend.{BackendConfigurationEntry, CromwellBackends}
+import cromwell.engine.workflow.WorkflowDescriptorBuilder
 import cromwell.engine.workflow.lifecycle.execution.WorkflowExecutionActor
 import cromwell.engine.workflow.lifecycle.execution.WorkflowExecutionActor.StartExecutingWorkflowCommand
 import cromwell.util.SampleWdl
